Add initial WebMangler core
Add an initial draft of the WebMangler core package, initializing the
repository in a monorepo format. The WebMangler core and related
packages will be written in TypeScript. The root contains a general
tsconfig.json file to support editors and running tests globally, as
well as a tsconfig.build.json file to be extended by individual package
tsconfig.json files for compilation and publishing.

This includes setting up ESLint, testing with Mocha [1], Chai [2], and
Sinon [3,4], test coverage with NYC [5], and Husky [6] for git hooks.

ESLint is configured with the following plugins:
- @typescript-eslint (of course)
- jsdoc [7] instead of tsdoc due to limited support
- security [8]
- mocha [9] & chai [10] for linting tests
- json [11] & yml [12] for linting configuration files

--
 1. https://mochajs.org/
 2. https://www.chaijs.com/
 3. https://sinonjs.org/
 4. https://www.npmjs.com/package/sinon-chai
 5. https://www.npmjs.com/package/nyc
 6. https://typicode.github.io/husky/#/
 7. https://www.npmjs.com/package/eslint-plugin-jsdoc
 8. https://www.npmjs.com/package/eslint-plugin-security
 9. https://www.npmjs.com/package/eslint-plugin-mocha
10. https://www.npmjs.com/package/eslint-plugin-chai-expect
11. https://www.npmjs.com/package/eslint-plugin-json
12. https://www.npmjs.com/package/eslint-plugin-yml