Update internal and exported type exports
Remove the "packages/" folder as a typeRoot in the root tsconfig.json
file (note that node_modules/@types is (one of) the default). And
update the core package to internally import and export types, as well
as properly export all relevant types in the NPM package. All types
previously defined in packages/core/index.d.ts are now defined
throughout packages/core/src in appropriate files, and the d.ts files
generated upon compilation will ensure these types are properly
distributed upon release. The `"esModuleInterop": false` option in the
root tsconfig.json ensures that the types are exported in the correct
way.