Initial WebMangler CLI package project folder
Add an initial draft of the WebMangler CLI package, adding the
project to the monorepo format. As per c55b57e, the WebMangler
CLI will be written in TypeScript, and has been initialized in
a similar fashion to the WebMangler Core.

In contrast to the WebMangler core, this module has yargs [1]
as a dependency, which will be used to parse CLI arguments.
Also, this project specifies the WebMangler core as a peer-
Dependency (and as a devDependency to be able to import it's
types during development).

--
1. https://yargs.js.org/