Add initial WebMangler CLI project source code
Add an initial draft of the WebMangler CLI that is capable
of running WebMangler core on one or more files with either
a default configuration, a configuration from a standard
configuration file (webmangler.config.js or .webmanglerrc
.js), or a configured configuration file (with the --config
option) AND is capable of writig the mangling back to the
input files.

Additionally, an override of the ESLint configuration for
package/cli was added, currently only to specify that the
security warning of reading arbitrary files (i.e security/
detect-non-literal-fs-filename) should be disabled for the
CLI project, as it wants to read arbitrary files to be
able to mangle them (or read configuration).