Update webmangler-core peerDep of webmangler-cli
Update the webmangler-core minimum peerDependency version on
webmangler-cli to v0.1.4 (as well as its listing as devDependency)
to make use of the shorter import paths introduced in v0.1.4. As
this is a breaking change (in alpha), the minimum peerDependency
version had to be increased.

The example in the README has been updated as well, and some minor
formatting issues in the README have also been fixed.