Add testdata directory
Add a simple testdate directory with simple CSS, HTML, and
JavaScript to test the WebMangler and specifically the CLI.
In the testdata directory, all files are ignored except the
ones that are put there intentionally, this would allow
anyone to create extra test files on demand without poluting
the directory.

Additionally, the packages/cli/.gitignore has been updated to
ignore the default configuration files or the CLI in its root,
preventing those from accidentally ending up in the commit
history if created for manual testing purposes.