chore: set up commitlint with conventional commits
Add commitlint [1] and at-commitlint/config-conventional [2] as root
dependencies to enable the usage of commitlint for commits to this
project. The commitlint.config.js file has been added to configure
commitlint, it extends te standard at-commitlint/config-conventional
configuration with project-specific scopes and some custom types, and
additionally uses shorter line lengths (72 instead of 100). (The commit
lint configuration has also been added to the ESLint configuration so
that it follows the correct style.) Commitlint will be used to verify
the commit message of every commit through the `commit-msg` git
hook [3] using Husky.

As an aside, the version field was re-added to package*.json to support
systems that expect it to be there - the NPM documentation doesn't
explicitly allow you to leave it out after all. I intend to use the
field as a general indication of where the monorope as a whole is at, so
now it uses the '-alpha' suffix to indicate that the project is still in
alpha state.

--
1. https://www.npmjs.com/package/commitlint
2. https://www.npmjs.com/package/@commitlint/config-conventional
3. https://git-scm.com/book/en/v2/Customizing-Git-Git-Hooks