feat(core): input files extend ManglerFiles
The definition of webmangler (the function exported by the webmanger
core package) is now in terms of any type extending ManglerFile. This
simplifies calling this function (in TypeScript) because it provides the
guarentee that files outputted by WebMangler are of the same type as
files inputted to the WebMangler - with only their contents changed.

This change required an update to the type definition of `ManglerEngine`
and `WebManglerPlugin`. The latter implies changes in the built-in
WebManglerPlugins.

Additionally, package/core has also been version bumped to v0.1.6 in
package*.json and the CHANGELOG.