test(cli): initialize test suite for fs adaptor
Initialize a test suite for the CLI fs adaptor module. The test suite
provides a 100% test coverage of the adaptor. To achieve this, three
test suites where created: one for the custom file class, one for
reading files, and one for writing files. Additionally, two mocks are
added to aid testing: one for the fs module and one for the custom file
class.

Additionally, the fs module was refactored slightly:

- Rename the `file.ts` file to `file.test.ts` for clarity and
  consistency.
- Split `index.ts` into `read.ts` and `write.ts`.
- `WebManglerCliFile` is now an exported as an interface.