feat(core): charSet as input to ManglerEngine
Add support to provide a custom character set to the ManglerEngine.
This can be done in the same object as is used to specify reserved names
and a mangler prefix. The custom character set will be used to generate
names for the output of mangling.

This will allow WebMangler plugins to specify custom character sets,
tailored to what they mangle. This allows for further reduction of file
sizes by the mangler as larger character sets mean more shorter strings.