feat(core): custom character set in SimpleMangler
Add support to specify a custom character set in the SimpleMangler
abstract class. This can be used by implementation to specify a
character set to be used for mangling - which may change depending on
what is being mangled.

The field for the character set is optional. If it is omitted the
default character set defined by the NameGenerator will be used.