feat(core): custom char sets for built-in manglers
Define and use custom character for the built-in manglers. The character
sets are are defined so that any permutation of the set forms a valid
string for the respective mangler. This, however, is still not optimal
for every mangler (except the HTML ID mangler) as some characters are
allowed in strings but not e.g. as first character. More specifically:

- CSS classes can contain hyphens, but not start with hyphens.
- CSS variables can contains hyphens and number, but not start with
  either.
- HTML attributes can contains hyphens and number, but not start with
  either.