feat(core): include hyphens in CSS class mangler
Add the hyphen to the character set of the CSS class mangler, allowing
it to generate class names with a hyphen. This relies on the new feature
which allows reserved names to be expressed as regular expression as CSS
class names are allowed to contain hyphens but not to start with
hyphens. Hence, the CSS class mangler will now ALWAYS list any class
starting with a hyphen (`/-.*/`) as "reserved" (regardless of the
options provided to the CSS class mangler).

The CSS Class mangler test suite has been updated to verify that mangled
class names never start with a hyphen. Both when no other reserved
strings are specified and when there are.