feat(core): include numbers & hyphens in CSS variable mangler
Add both numbers and the hyphen to the character set of the CSS variable
mangler, allowing it to generate variable names with both numbers and
hyphen. This relies on the new feature which allows reserved names to be
expressed as regular expression as CSS variable names are allowed to
contain hyphens and numbers but not start with hyphens or numbers.
Hence, the CSS variable mangler will now ALWAYS list any class starting
with a hyphen (`/-.*/`) as "reserved" (regardless of the options
provided to the CSS variable mangler).

The CSS variabl emangler test suite has been updated to verify that
mangled class names never start with a hyphen or number. Both when no
other reserved strings are specified and when there are.

Additionally, the same functionality and tests for the CSS class mangler
has been updated to be in line with the appraoch taken for the CSS
variable mangler. Most notable, the tests for illegal class names is now
more focused.