feat(core): include numbers, "_", and "-" in HTML attribute mangler
Add both numbers, underscores, and hyphens to the character set of the
HTML  attribute mangler, allowing it to generate attribute names with
numbers, underscores, and hyphens. This relies on the new feature which
allows  reserved names to be expressed as regular expression as HTML
attribute names are allowed to contain numbers, underscores, and hyphens
but not start with any of them. Hence, the HTML attribute mangler will
now ALWAYS list any attribute name starting with a number, underscore,
or hyphen as reserved.

The HTML attribute mangler test suite has been updated to verify that
mangled attribute names never start with a number, underscore or hyphen.
Both when no other reserved strings are specified and when there are.

Additionally, the same tests for both the CSS class mangler and CSS
variable mangler have been updated with some minor fixes.