fix(core): bug where prefix could not be omitted for HTML attributes
Fix a bug where prefixes could not be omitted with the HTML attribute
mangler. This problem arrised from the way the `getPrefix` helper method
was written.

For safety, all similar helper functions in every built-in mangler are
rewritten to avoid similar issues. However, I confirmed that none of the
other manglers was affected by this kind of bug as the default value was
always the value that would cause the bug.