test: create, publish, and use testing utilities library
Create a new package called "testing" that will be used to define
testing utilities. Currently, it defines the TestScenario type that was
previously defined in several places throughout the core package. This
new package has been published and is used in core to replace every
local `TestScenario` definition. To achieve this, the `TestScenario`
type is generic in the TestCase type. Hence, every test suite using
`TestScenario` defines its own `TestCase` type for local usage.

Additionally, every `TestCase` type definition has been moved in front
of the test suite's first `suite` call so that everyone sees it so
they have context about how the tests in this file are defined.