feat(cli): create logger module
Create a new, simple logger module for the CLI package that defines a
basic `Logger` interface and provides a simple `DefaultLogger`
implementation. The package has a 100% test coverage.

This new package is used to support the new CLI option "--verbose" (or
"-v") for short. This is a boolean/count option that can be used to
tweak the amount of logging the CLI does. Note that as of this commit
stats are always outputted.