fix(core): maximum call stack size exceeded when attribute mangling
Fix the _Maximum call stack size exceeded_ error raised the HTML
Attribute mangler when it is used on large files. This is achieved
by reducing the number of capturing groups in the main pattern template.

Additionally, some refactoring related to this occured, removing now
unused or non-duplicate constants; and the test suite for the HTML
attribute mangler on HTML source code was improved.