refactor(core): add replaceAll method to ManglerExpression interface
Add a new method called `replaceAll` to the ManglerExpression interface
and deprecate the existing `replace` method. The new method is similar
to the now-deprecated `replace` method but instead of replacing one
string by another, it can be used to replace many strings in one call.
This allows for expressions-based optimizations w.r.t. the number of
replacements. This can be seen from the changes to the ManglerEngine,
which now has one call to `replaceAll` instead of n - the number of
strings to mangle - calls to `replace`. This is significant because the
number of strings to mangle can grow very large (whereas the number of
files and fileExpressions are, relatively, limited).

The gains that can be achieved by this can be seen from the new
`ParallelManglerExpression` class, which performs `replaceAll` in
"parallel", letting the Regular Expression engine do the heavy lifting
of search for all n replacements in one pass over the document. This
gives significant performances boosts 1) the Regular Expression engine
can be optimized far better than any JavaScript code and 2) the engine
won't go through each file n times (which can be significant for large
files).

The new `ParallelManglerExpression` class is currently used by the HTML
ID mangler's HTML support.

The existing `_ManglerExpressions` class has been deprecated due to
plans to move it to the same package as the new `ManglerExpression`
implementation.