perf(core): improve CSS class mangling performance in HTML
Create a new ManglerExpression implementation, used by the CSS class
mangler's HTML language support, that works in two stages. For CSS
classes, it first finds any class attribute and fetches its value iff
the value matches the pattern of interest at least once. Then, it
matches the value substring against the pattern of interest and returns
or replaces all instances of the pattern in the substring.