refactor(core): deprecate WebManglerPlugin.use
Deprecate the `use` method of the `WebManglerPlugin` type as it's
implementation would always be the same: collection all
ManglerExpressions from each provided WebManglerLanguagePlugin and
return them all when `config` is called. Instead, the WebMangler core
will now take care of collecting the ManglerExpressions from every
specified WebManglerLanguagePlugin and provide them to the MangleEngine
separately.

This has the added benefits of WebManglerPlugins and
WebManglerLanguagePlugins becoming completely decoupled and removing
the need fr the abstract helper class `BaseManglerPlugin`.