test(core): update all mangler test suites for CSS
Update and improve the test suites for all manglers for CSS. This
includes refactoring the test source for readability and
maintainability, adding new tests, and improving existing tests.

A couple local lists of values where moved to the `css-constants.ts`
helper. The `test-helpers.ts` utility now has an `isValidAttributeName`
method (similar to the similarly named function for classes and IDs,
whose documentation has been fixed).

Also, increase the timeout for running tests with Mocha to 5 seconds. A
couple of the scenarios would occasionally exceed 2 seconds.