refactor(core): make WebManglerLanguagePlugin more generic
Update the WebManglerLanguagePlugin interface with a new method,
`getExpressionsFor`, can be called with a "name" and an "options object.
The "name" is used to indicate the "name" of the group of expressions
that should be returned (e.g. expressions that find attributes or query
selectors). The "options" can be used to configure the group of
expressions (e.g. only find attribute values of "id" and "for"
attributes). The builtin language plugins are updated to the extend
necessary to satisfy the interface AND to support mangling CSS classes,
HTLM attributes, and HTML IDs in only CSS and HTML.

The WebManglerPlugin interface is updated accordingly to be able to
retreive the "name"s and "options" they want to use. To this end, the
`.config` method has been deprecated (along with the `id` field on the
`MangleEngineOptions`) and is replaced by the new `.options` method,
which should return an object that is a `ManglerEngineOptions` instance
with an additional field that is a list of "name"s and "options" for
expressions - called `MangleOptions`. The builtin WebManglerPlugins have
been updated to the extend necessary to satisfy the interface AND to
support mangling CSS classes, HTLM attributes, and HTML IDs in only CSS
and HTML.

This commits omits in-code documentation except for at-version tags and
the bare minimum to pass the linter. This commit also omits updating
tests except to the extend necessary to be able to run the tests.