perf(core): reduce time spend mapping to unique values
Update the WebMangler core engine to, in terms of old code, reduce the
size of the `mapToUnique` map, thereby improving the perforance of the
mangle engine.

This is achieved by only mangling values to a unique intermediary value
if necessary (in many cases this wil be zero).

Additionally, ineffective mappings (e.g. mapping "a" to "a") are now
filtered out as well for a minor performance improvement.