fix(core): separate expressions properly by language
A bug in `getExpressions` in index.js and the WebMangleLangaugePlugin
caused incorrect separation of MangleExpressions by langauge, causes
expressions to be used for different languages than they are intended.
Both index.js and the WebManglerLanguagePlugin interface are fixed to
solve this problem by using the old approach of returning a mapping
instead of relying on the return value of the languages array (as that
doesn't work if the WebManglerLanguagePlugin provides support for
multiple languages).

This also fixes a minor "bug" with regards to unexpected mangled HTML by
the CSS variable mangler in HTML (see it's test suite for more details).