feat(core): any prefix & suffix for query-selectors expression set
Provide support for the query-selectors MangleExpression set to be
configured with any prefix, and also with a suffix. This may seem
unnecessary as most query selectors are either withing a well-defined
scope (e.g. classes and IDs) or should probably not be mangled (e.g.
elements). However, to allow for absolute freedom I decided to just
allow any prefix and suffix and trust developers to use them correctly -
adding a note to the options documentation as a minimum. One important
thing this would support is matching attribute selectors if needed.