feat(core): configure id-like attributes for built-in id mangler
Add a configuration option to the built-in HtmlIdMangler that allows
users to specify what HTML attributes should be threaded as IDs in
addition to the `id` and `for` attributes.

This option may be useful if id's are used as references in other (data)
attributes.

The option is tested in the HtmlIdMangler test suite. Including edge
cases such as an empty list or the inclusion of the always included `id`
and `for` attributes.