feat(core): configure href-like attriutes for built-in ID mangler
Add a configuration option to the built-in HtmlIdMangler that allows
users to specify what HTML attributes should be treated as `href`s in
addition to the `href` attribute itself.

This option may be useful if hrefs are specified separately, or if this
implementation missed a standard href-like attribute.

The option is tested in the HtmlIdMangler test suite, including edge
cases such as en empty list or the inclusion of the always included
`href` attribute.