test(core): setup benchmarking of language plugins
Setup benchmarking for the builtin language plugins through Mocha. To
achieve this, the Mocha configuration file has been changed to a
JavaScript file to dynamically change the configuration based on e.g.
environment variables. This is used in the root package.json to add a
new script called "benchmark" that starts Mocha with the TEST_ENV
environment variable set to "benchmark".

A sample of 3 benchmarking tests has been added, along with a helpers
file that contains the logic for benchmarking functions. Currently,
only the function execution time is benchmarked and it is checked that
they don't exceed a certain limit set based on local experiments.