fix(core): id mangling in URLs with query parameters
Fix a bug in the HtmlIdMangler where URLs with query parameters would
not be mangled due to a bug w.r.t. the structure of URLs. The existing
implementation expected query parameters AFTER the id fragment whereas
in reality they come before the id fragment (see e.g. [1]).

The HtmlIdMangler test suite has been updated with a test case for URLs
with query parameters to verify the correct behaviour.

--
1. https://en.wikipedia.org/wiki/Uniform_Resource_Identifier