chore: remove max-len exception of 120 characters for test code
The exception was intended to allow for longer test descriptions, which
is what it is rarely used for. Hence, this commit removes the exception
from both the ESLint and EditorConfig configuration.

It also broadens the set of files considerd as "test" files, this may
become useful if non-mocha rules are applied to tests (such as disabling
security rules).