refactor(core): remove v0.1.15 deprecated fields and constructors
Remove deprecated statis fields from built-in manglers and replace the
`SimpleManglerPlugin` constructor for one without an identifier.

Update the test suites for the built-in manglers and the
`SimpleManglerPlugin` accordingly. Additionally, the configuration
tests for the built-in manglers are now structured better and no
longer use incorrect variable names.

The built-in manglers now relay on a character set of all characters
provided (new) by the characters module to test for illegal names
being generated. The idea being that any character set is at most
as big as all characters, so relying on the size of the character set
of all characters should be sufficient in all cases.

Lastly, some remaining `.map().flat()` chains are replaced by
`.flatMap()` in the built-in manglers test suites.