benchmark(core): test files changed and benchmark unchanged files
This commit introduces benchmarking of the language plugins against
files that don't contain any match. This is motivated by problems solved
in e.g. [1] where some manglers are incredibly slow if there are no (or
very few) matches in a file.

Additionally, this adds verification to the existing benchmarks that
something in the file actually changed. To achieve this, the structure
of individual benchmark tests was updated as well.

--
1. 75918bf17d70cc7891c5fc21ba14c41f5cf33b32