feat(testing): add mocks for three more WebMangler interfaces
Add three more mocks to the testing utilities package for the
MangleExpression interface, the WebManglerLanguagePlugin interface, and
the WebManglerPlugin interface. The latter two are based on the existing
mocks present in package/core. The former isn't, as the existing mock
contains more logic than is really necessary, instead this mock is
created in the same style as the latter two.

These mocks use SinonJS [1] in their implementation, hence Sinon is
added as a peerDependency to the package/testing/package.json.

--
1. https://sinonjs.org/