refactor(cli): improve fs module
Refactor the internal fs module for the CLI package by splitting based
on concerns (particularly extracting the logic for listing files
recursively) and overal improvements to style and naming, as well as
improvements to the test suites (primarily due to the use of the
TestScenario type).

This refactoring includes changes that make reading and writing multiple
files in parallel. This changes has been included in the CHANGELOG.

Also, the example in the README has been improved by using the
`RecommendedManglers` instead of `BuiltInManglers`.