feat(core): support unqouted HTML attribute values
As per the W3C specifications [1], this adds support or unquoted
attribute-value syntax in HTML. Specifically, the
"single-value-attribute" category of expressions now supports unquoted
attribute values.

The implementation here is a bit more forgiving than the specs. E.g. it
does not require a space between the value and the "/" of a self-closing
tag.

--
1. https://www.w3.org/TR/2012/WD-html-markup-20120329/syntax.html