chore(core): remove now-unused attribute selector attributes sets
Remove the expressions for attribute selectors in the "attributes" set
of expressions of the built-in CSS & JS language plugins. In the case
of the JavaScript langauge plugins, this entails removing the
attributes expression factory entirely.

These expressions used to be used by the HTML attribute mangler, but
that mangler has been refactored to achieve this through the
"query-selectors" set of expressions instead.