chore: refine linting rules for MarkDown
Add additionaly MarkDown linting rules from [1] that are not incuded in
eslint-plugin-md/remark-preset-lint-markdown-style-guide [2]. Also,
explicitly disable the rules omitted intentionally in [3]. Lastly,
disable the "lint-maximum-line-length" rule and rely on the standard
"max-len" rule instead.

Unfortunately, the lint-no-missing-blank-lines rule [4] could not be
added as it reports false positives for code blocks.

--
1. https://github.com/remarkjs/remark-lint/blob/main/doc/rules.md
2. https://github.com/remarkjs/remark-lint/tree/main/packages/
  remark-preset-lint-markdown-style-guide
3. 3ee726182279fdbd6180a7cfd782ecc9ecd6361a
4. https://github.com/remarkjs/remark-lint/tree/main/packages/
  remark-lint-no-missing-blank-lines#use