ci: configure Codecov for package separated coverage reports
Update the Codecov configuration, following [1], to handle the new
package separated testing approach.

The carryforward setting [2] ensures coverage is carried through commits
in the case that commit does not receive coverage for the project. This
would support only running tests for the package that was actually
touched in a commit.

--
1. https://docs.codecov.io/docs/flags
2. https://docs.codecov.io/docs/flags#carryforward-flags