feat(core): add ability to configure the BuiltInLanguagesPlugin
Add an optional parameter to the BuiltInLanguagesPlugin that allows for
configuring the individual language plugins it consists of.

In contrast to the multi-mangler plugins, I did not add the ability to
disable any of the individual language plugins. Primarily because it
makes less sense to disable a language plugin than a mangler. A mangler
might be malfunctioning or unwanted. Disabling a malfunctioning language
plugin does not fix anything, and unwanted/unneeded language plugins are
simply not used. That said, it _could_ always be added later.