benchmark(core): improve JavaScript language plugin benchmarking
Improve the benchmarking of the JS language plugin by making the text
used for benchmarking more like real-life and by including pieces of
valid but perhaps uncommon JavaScript that are known to cause slow
mangling. Specifically long strings.

Additionally, this reduces the runtime budget for the query selector
mangler expressions to the standard 10ms and updates the query selector
mangler epxressions so that it's performance is below that.