chore(core): update CHANGELOG
Add performance increase as a result of improved benchmarking to the
CHANGELOG.

Also reduce the budget of all small file benchmarks from 0.5 to 0.3,
of large files (with something to mangle) from 10 to 6, and large
files (without something to mangle) from 10 to 1. Making these numbers
smallers makes the benchmarks more effective at keeping the core fast.
The numbers where chosen based on running the benchmarks a couple of
times and seeing the curren results.