test(core): cass variable in unquoted style values
Adds a test scenario for each test instance of CSS variables in HTML
for unquoted style attribute values.

Because this is only supposed to work if the style value has NO spaces,
the test instances for varied spacing had to be replaced by test
scenarios that vary the spacing for each test instance.