chore(core): increase HTML multi-value attribute mangling runtime budget
This is one of the most complex expressions and a recent change to
support unquoted values has dubbeled the number of expressions making it
about twice as slow. It's runtime budget was increased to 10 ms as 1)
this is still considered a reasonable budget and 2) no immediately
obvious improvements can be made to the expression.