test(core): include double counted findings in language tests
Some of the expression factories of the built-in language plugins report
some of their findings twice. Even though this is not intended
behaviour(*), the expectation is currently set accordingly. This is
because fixing this is not trivial and so this should be done
separately.

(*): The reason this is not desired is because mangled values are
determined based on the frequency of the strings to mangled.