fix(core): match nested query selectors
Fix the bug where nested query selectors (e.g. in a media query) would
not be found nor replaced in CSS. This has been fixed by removing part
of the regular expression (that used) to ensure the query selector was
not present in between curly braces ("{" and "}").

Unfortunately, this introduces a new bug where query selectors may now
be found and replaced in CSS strings...