chore: update builds to use local dependencies
Update the build systems for all packages so that internal dependencies,
such as core depending on at-webmangler/testing, can be truely local
rather then through NPM. The main motivation for this is to streamline
cross-package development, see for example (merge) commit:

    accfa0bca67a03656015f13f726b3fa2b5515d3f

In practice, this means all dependency that can be local have been
removed from package*.json files, all packages that are depended upon
locally now have "composite" set to true in their tsconfig.json [1], and
the packages depending on them reference the ones they need through the
"references" field of the tsconfig.json. Additionally, the root
tsconfig.json now defines (alias) "paths" [2] for local imports such
that they can still be imported as if they're absolute.

The above changes are sufficient for compiling all packages. The only
drawback with the current approach is that local devDependencies will
also be compiled when this is unnecessary.

Continueing on devDependencies, as tests are run without compilation the
above is not sufficient for running test suites. For that to work I
added _tsconfig-paths_ [3], a ts-node [4] plugin that resolves the paths
[2] for you. This is referenced in .mochars.js. Additionally, to keep it
simple to run packages/cli without compilation, a script with ts-node
and tsconfig-paths has been added.

Additional changes:
- All packages now compile into a folder named "build". The npm manifest
  and .gitignore for each project has been updated to reflect that. This
  includes the usage of "exports" [5] in packages/cores npm manifest, it
  ensures the languages and manglers (and their utils) can be imported
  without having to import, e.g., "webmangler/build/languages". This
  feature ([5]) has been present since Node v12 so using it should be
  safe.
- Project .npmignore files have been normalized and updated to ignore
  new files - in particular the `tsconfig.tsbuildinfo` files generated
  by TypeScript now that "references" [1] are used.
- The root .gitignore file has been updated to ignore files generated
  by TypeScript due to the use of "references" [1].
- The clean script has been updated to clean the benchmarking package.
- Update the push-checks.yml workflow to compile before running tests
  and benchmarks. This is needed because tests and benchmarks now use
  local dependencies, which need to be compiled before they can be used.

--
1. https://www.typescriptlang.org/docs/handbook/project-references.html
2. https://www.typescriptlang.org/tsconfig#paths
3. https://www.npmjs.com/package/tsconfig-paths
4. https://www.npmjs.com/package/ts-node