refactor(core): remove ignoreStrings from SingleGroupMangleExpression
Disable the recently introduced `ignoreStrings` constructor parameter
from the SingleGroupMangleExpression utility. Instead, this should be
implemented into the `patternTemplate`. This was not possible before
because the class would error if no match was found, whereas the new
behaviour is such that a match is ignored if the specified group is not
found. Hence, the wrapping group implemented by `prepareTemplate` can
simply be done by the user of this utility - as seen in the updated
`newCssSelectorExpression` of the built-in CSS language plugin.