refactor(core): add support for mangling embedded code
Add support mangling embedded code, e.g. CSS or JavaScript embedded in
HTML. This is achieved by extending the WebManglerLanguagePlugin type
with a method to `getEmbeds`. This method must return zero or more
so-called `WebManglerEmbed`s for a given `WebMangerFile`. A
`WebManglerEmbed` is a new type that extends `WebMangerFile` and
additionally includes a start and end index of where the embed appears
in the origin file. On top of that it must have a method to obtain the
"raw" embed. This is needed because the embed must represent a valid
`WebManglerFile`, even though the raw embed is not. Take for example the
CSS in a the style attribute in HTML:

  <div style="color: red;"></div>

The content of the style attribute by itself is not valid CSS, but it is
if it is wrapped into a selector block, e.g.:

  div { color: red; }

So, that is what the `content` of the WebManglerEmbed *may* be set to,
but the `rawEmbed` is still "color: red;".

The core's entry function has been updated to extract embeds from the
provided files, mangle the files as well as the embeds, and re-embed the
embeds back into the original files.

The new embeds.ts module houses the logic for systematically extracting
embeds from files and re-embedding them after mangling. Note that it
replaces embedded code by uniquely identifiable strings in the file's
`content` before mangling begins. This is to ensure the embedded code
can be found and replaced after mangling. This must be done based on the
`startIndex` and `endIndex` of the WebManglerEmbed because, e.g.
embedded CSS may appear as content in a HTML page.

The built-in WebManglerLanguagePlugin helpers have been updated to
support the `getEmbeds` method. The SimpleLanguagePlugin defaults to
returning no embeds and can be configured to return embeds by providing
one or more so-called `EmbedGetters`, which are basically just
implementations for `getEmbeds`. The MultiLanguagePlugin simply returns
all embeds of all plugins it contains.

Lastly, the WebManglerLangaugePluginMock from the testing packages has
been updated to include a `getEmbeds` stub.