refactor(core): remove CSS properties & values support from HTML plugin
Remove the custom support for finding and replacing CSS properties and
CSS values from the built-in HTML language plugin and instead rely on
the new embed functionality to find style attributes and mangle CSS
properties and CSS values in them.

In order for this to work - i.e. to pass the existing test suites - the
built-in HTML language plugin's style attribute embeds finder has been
updated for the following:

- Attribute values with ">"
- Multiple style attributes on an element
- Empty style attribute values
- Unquoted style attribute values