chore(core,cli): support running tests in parallel with Mocha
Update the test suite to support running the tests in parallel using
Mocha [1]. This feature is not, and will not, be enabled by default. For
one, it should never be used when running benchmark tests because they
are time-sensitive. Secondly, running tests in parallel has both upsides
and downsides - I found [2] to be a good read on this for NodeJS - so
enabling it may lead to worse performance. On my machine, I found that
tests run ~4 seconds faster.

This existing source and tests largely avoid any problems that may arise
due to the use of parallel tests. However, there were some issues that
have been fixed.

- All test files that use sinon-chai now initialize it correctly. This
  was not a problem before because it would be loaded as a result of
  other tests running before it.
- All mocks now have only instance-specific properties. This prevents
  any issues that may arise from multiple modules sharing state.
- Module mocking is now done using proxyquire [3]. I found that mocking
  imports, specifically default imports, using Sinon sometimes caused
  problems. Whereas proxyquire works just fine. In the long run, it may
  be worth investigating if proxyquire can be replaced with dependency
  injection.

--
1. https://mochajs.org/#parallel-tests
2. https://developer.ibm.com/languages/node-js/articles/parallel-tests-mocha-v8/
3. https://www.npmjs.com/package/proxyquire