docs: remove leading '$' in shell snippets in markdown
Remove the leading '$' used in shell snippets in MarkDown files to
indicate this is a shell command. This makes it easier to copy the
command, especially if the generated HTML provides a copy option for the
snippet. Instead, the text leading up to the snippet should make clear
it is a shell command, as can the formatting and text content itself.