chore(core,testing): version bump to 0.1.21 and 0.1.4 respectively
Bump the version of package/core in its manifest and changelog.
Bump the version of package/testing in its manifest and changelog.

Additionally, fix a typo in the README of package/core and add the
`getEmbeds` mock to the documentation of the WebManglerLanguagePlugin
mock in the README of package/testing.