test(core): generate cases for CSS language plugin's selector factory
Convert the test suite for the CSS language plugin's query selector
factory to be based on generated test cases using predefined value
presets. This includes creating a value sample of CSS declarations and
creating a value preset for both before and after query selectors as
well as CSS declarations.

Additionally. improve some of the test-generation helper methods and the
generated test suites for the CSS property and CSS value expression
factories of the CSS language plugin. This includes the removal of the
generic TestCase type for the CSS language plugin expression factory
tests as it turns out it isn't generic enough.

Lastly, the test generation for query selectors does not properly cover
nested query selectors yet, the current scenario for that is abusing the
generation system. Further effort is needed to find an "appropriate"
solution.